Update README.md
Just learned that the docker command called build has to have a dot at the end of the command. The -t arguement allows you to name your build however you want. I decided not to include it because there are sometimes people out there that are willing to put in a different build name and that's fine.